Improve symlink traversal of mingw stat and chdir
Based on resolve_symlink in lockfile.c, this resolves paths when the path along
the way has symlinks.

It uses a left-to-right approach, checking each part of the path
for a symlink and then replacing that section (depending on whether it is
absolute or relative) with the link.

This is required for make_absolute_path tests to succeed.

Signed-off-by: Michael Geddes <michael@frog.wheelycreek.net>