blame: Make sure crlf conversion happens using diff textconv interface
The whole crlf conversion seemed to be missing; just reading in the whole file
unless there was an actual textconv.

There were also issues with memory lifetime.  The free_filespec was deleting
the memory being used, so it needed to be detached properly, taking into
account that an xalloc'd piece of memory is wanted on detach and not static
memory or a mmap'd file.

Signed-off-by: Michael Geddes <michael@frog.wheelycreek.net>